+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 * DEPENDENCIES
   * Updated ReverseDSC to version 2.0.0.13;
   FIXES [#2270](https://github.com/microsoft/Microsoft365DSC/issues/2270)
+* MISC
+  * Added support for Exchange to the Update-M365DSCAzureAdApplication cmdlet
 
 # 1.22.1005.1